modularize some of the within epoch training
